+++ packages/client-main/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/bsorrentino/js-notebook/compare/v0.4.10...v0.5.0) (2022-03-14)
+
+**Note:** Version bump only for package @bsorrentino/jsnotebook-client-main
+
+
+
+
+
 # [0.4.0](https://github.com/bsorrentino/js-notebook/compare/v0.2.3...v0.4.0) (2022-03-13)
 
 **Note:** Version bump only for package @bsorrentino/jsnotebook-client-main