+++ packages/local-pkg/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/bsorrentino/js-notebook/compare/v0.4.10...v0.5.0) (2022-03-14)
+
+
+### Features
+
+* **server:** allow configuration of static routes ([83fa6c4](https://github.com/bsorrentino/js-notebook/commit/83fa6c4c21c38b2321091316790e04cce8677ce1)), closes [#7](https://github.com/bsorrentino/js-notebook/issues/7)
+
+
+
+
+
 # [0.4.0](https://github.com/bsorrentino/js-notebook/compare/v0.2.3...v0.4.0) (2022-03-13)
 
 **Note:** Version bump only for package @bsorrentino/jsnotebook-local-pkg